MB-25422: Add getMeta to default_engine for tests
This patch implements the getMeta opcode for the default_engine so that
one can find the datatype of a deleted document in memcached_testapp
tests.

While the getMeta functionality could be moved in to memcached, we
haven't done it at this time as it would disrupt the regular ep-engine
getMeta code path. Which is not something we want to do for spock. The
task to move this functionality in to memcached is tracked by MB-25427.

Change-Id: I1c2ded2cfb38736d9f1c54a2c12d6d0841541f63